requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
black==24.10.0
boto3==1.34.21
botocore==1.34.142
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap5==2023.10
cryptography==42.0.5
defusedxml==0.8.0rc2
distro==1.9.0
dj-database-url==2.2.0
Django==5.1.2
django-crispy-forms==2.1
django-storages==1.14.2
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
flake8==7.1.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.4
httpx==0.26.0
idna==3.7
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
openai==1.23.6
packaging==23.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.6
psycopg2==2.9.9
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2021.3
requests==2.32.2
requests-oauthlib==1.3.1
s3transfer==0.10.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.9.0
urllib3==2.2.2
whitenoise==6.6.0
zope.interface==6.1
